AsyncScrolling: add ScrollingTreeFixedNode base class https://bugs.webkit.org/show_bug.cgi?id=246612
Reviewed by Žan Doberšek.

Make Cocoa and Nicosia implementations inherit from the base class and move the common code there.

* Source/WebCore/PlatformMac.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/SourcesCocoa.txt:
* Source/WebCore/SourcesGTK.txt:
* Source/WebCore/SourcesWPE.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/page/scrolling/ScrollingTreeFixedNode.cpp: Added.
(WebCore::ScrollingTreeFixedNode::ScrollingTreeFixedNode):
(WebCore::ScrollingTreeFixedNode::~ScrollingTreeFixedNode):
(WebCore::ScrollingTreeFixedNode::commitStateBeforeChildren):
(WebCore::ScrollingTreeFixedNode::computeLayerPosition const):
(WebCore::ScrollingTreeFixedNode::dumpProperties const):
* Source/WebCore/page/scrolling/ScrollingTreeFixedNode.h: Copied from Source/WebCore/page/scrolling/cocoa/ScrollingTreeFixedNode.h.
* Source/WebCore/page/scrolling/ScrollingTreeNode.h:
(WebCore::ScrollingTreeNode::isFixedNodeCocoa const):
(WebCore::ScrollingTreeNode::isFixedNodeNicosia const):
* Source/WebCore/page/scrolling/ScrollingTreePositionedNode.h:
* Source/WebCore/page/scrolling/ScrollingTreeStickyNode.h:
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeFixedNode.mm: Removed.
(WebCore::ScrollingTreeFixedNode::create): Deleted.
(WebCore::ScrollingTreeFixedNode::ScrollingTreeFixedNode): Deleted.
(WebCore::ScrollingTreeFixedNode::~ScrollingTreeFixedNode): Deleted.
(WebCore::ScrollingTreeFixedNode::commitStateBeforeChildren): Deleted.
(WebCore::ScrollingTreeFixedNode::applyLayerPositions): Deleted.
(WebCore::ScrollingTreeFixedNode::dumpProperties const): Deleted.
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeFixedNodeCocoa.h: Renamed from Source/WebCore/page/scrolling/cocoa/ScrollingTreeFixedNode.h.
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeFixedNodeCocoa.mm: Added.
(WebCore::ScrollingTreeFixedNodeCocoa::create):
(WebCore::ScrollingTreeFixedNodeCocoa::ScrollingTreeFixedNodeCocoa):
(WebCore::ScrollingTreeFixedNodeCocoa::commitStateBeforeChildren):
(WebCore::ScrollingTreeFixedNodeCocoa::applyLayerPositions):
(WebCore::ScrollingTreeFixedNodeCocoa::dumpProperties const):
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNode.mm:
* Source/WebCore/page/scrolling/cocoa/ScrollingTreePositionedNodeCocoa.h:
(WebCore::ScrollingTreePositionedNodeCocoa::layer const): Deleted.
* Source/WebCore/page/scrolling/cocoa/ScrollingTreeStickyNodeCocoa.h:
(WebCore::ScrollingTreeStickyNodeCocoa::layer): Deleted.
* Source/WebCore/page/scrolling/mac/ScrollingCoordinatorMac.mm:
* Source/WebCore/page/scrolling/mac/ScrollingTreeMac.mm:
(ScrollingTreeMac::createScrollingTreeNode):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNode.cpp: Removed.
(WebCore::ScrollingTreeFixedNode::create): Deleted.
(WebCore::ScrollingTreeFixedNode::ScrollingTreeFixedNode): Deleted.
(WebCore::ScrollingTreeFixedNode::~ScrollingTreeFixedNode): Deleted.
(WebCore::ScrollingTreeFixedNode::commitStateBeforeChildren): Deleted.
(WebCore::ScrollingTreeFixedNode::applyLayerPositions): Deleted.
(WebCore::ScrollingTreeFixedNode::dumpProperties const): Deleted.
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.cpp: Added.
(WebCore::ScrollingTreeFixedNodeNicosia::create):
(WebCore::ScrollingTreeFixedNodeNicosia::ScrollingTreeFixedNodeNicosia):
(WebCore::ScrollingTreeFixedNodeNicosia::commitStateBeforeChildren):
(WebCore::ScrollingTreeFixedNodeNicosia::applyLayerPositions):
(WebCore::ScrollingTreeFixedNodeNicosia::dumpProperties const):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.h: Renamed from Source/WebCore/page/scrolling/nicosia/ScrollingTreeFixedNode.h.
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeNicosia.cpp:
(WebCore::ScrollingTreeNicosia::createScrollingTreeNode):
* Source/WebCore/page/scrolling/nicosia/ScrollingTreeStickyNodeNicosia.h:
* Source/WebCore/platform/TextureMapper.cmake:
* Source/WebKit/UIProcess/RemoteLayerTree/RemoteScrollingTree.cpp:
(WebKit::RemoteScrollingTree::createScrollingTreeNode):
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/RemoteScrollingCoordinator.mm:

Canonical link: https://commits.webkit.org/255731@main
