+++ Source/WebCore/PlatformMac.cmake
@@ -555,7 +555,7 @@ list(APPEND WebCore_PRIVATE_FRAMEWORK_HEADERS
 
     page/scrolling/ScrollingStateOverflowScrollProxyNode.h
 
-    page/scrolling/cocoa/ScrollingTreeFixedNode.h
+    page/scrolling/cocoa/ScrollingTreeFixedNodeCocoa.h
     page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNode.h
     page/scrolling/cocoa/ScrollingTreePositionedNode.h
     page/scrolling/cocoa/ScrollingTreeStickyNodeCocoa.h
