+++ Source/WebCore/SourcesWPE.txt
@@ -48,7 +48,7 @@ page/linux/ResourceUsageThreadLinux.cpp
 
 page/scrolling/nicosia/ScrollingCoordinatorNicosia.cpp
 page/scrolling/nicosia/ScrollingStateNodeNicosia.cpp
-page/scrolling/nicosia/ScrollingTreeFixedNode.cpp
+page/scrolling/nicosia/ScrollingTreeFixedNodeNicosia.cpp
 page/scrolling/nicosia/ScrollingTreeFrameScrollingNodeNicosia.cpp
 page/scrolling/nicosia/ScrollingTreeNicosia.cpp
 page/scrolling/nicosia/ScrollingTreeOverflowScrollProxyNode.cpp
