+++ Source/WebCore/page/scrolling/cocoa/ScrollingTreeOverflowScrollProxyNode.mm
@@ -31,6 +31,7 @@
 #import "ScrollingStateOverflowScrollProxyNode.h"
 #import "ScrollingStateTree.h"
 #import "ScrollingTree.h"
+#import "ScrollingTreeOverflowScrollingNode.h"
 #import "WebCoreCALayerExtras.h"
 
 namespace WebCore {
